sql/parser: move the ColumnType converters to their own file
